+++ componenti_pc.h
@@ -8,7 +8,7 @@ private:
     unsigned int altezza;
     std::string nome;
     std::string produttore;
-    float prezzo;
+    double prezzo;
     unsigned int consumo_energetico;
 public:
     Componenti_PC(unsigned int =0, unsigned int =0, std::string ="Non specificato", std::string ="Non specificato", float =0, unsigned int =0);
@@ -22,8 +22,8 @@ public:
     void setNome(std::string);
     std::string getProduttore() const;
     void setProduttore(std::string);
-    float getPrezzo() const;
-    void setPrezzo(float);
+    double getPrezzo() const;
+    void setPrezzo(double);
     unsigned int getConsumoEnergetico() const;
     void setConsumoEnergetico(unsigned int);
     virtual ~Componenti_PC() =default;
